Pass array of new prefferences
Preferences keep track of their own index in the prefereces array,
when a preference is changed, the new value is assigned via the preference
index.
Now we are hitting a bug, the new array doesn't contain a venue 'key'
and thus doesn't get merged
by the reducer. Using Object.assign copies the device key, but the object is
no longer an Array. :/