Add blocks B2 and C2, turn on debug
Why: Recieved bug report from Facebook about crashes, discovered that it only happened when connecting to devices with block B2 installed. found that I didn't have a defenition for block B2 in the program at all.

*---

How: working with Ross in Test Flight, so turned debug on and uploading a test build

*---