Debug IportData picker views
Why: There were some unfinished modifications to the way this component was built
I seem to remembering copying the basic code from a prototype I had built. It seems
I never finished implementing some of the state management.

*---

How: Resolved syntax and reference errors. The "data" variable
was undefined. I intend to eventually connect this compoent to the Redux state
but for testing this will work for now

*---